clean up helpers.jl

+++ helpers.jl
@@ -15,14 +15,24 @@ function get_input(year::Integer, day::Integer)
         "session" => "53616c7465645f5f6da2467223e04715bff79670b7e5af29336601520c92a0df6322bae9ab34e94a0c8b9df9e791557027b6e031629fbfb911fb7e38b2280682",
     )
 
-    response = HTTP.request("GET", advent_url, cookies = cookies)
-    input = String(response.body)
-
     file = "inputs/input$day" * "_$year"
 
-    open(file, "w") do fp
-        write(fp, input)
-        println(input)
+    if isfile(file)
+        input = read(file, String)
+    else
+        response = HTTP.request("GET", advent_url, cookies = cookies)
+        input = String(response.body)
+
+        open(file, "w") do fp
+            write(fp, input)
+        end
+    end
+
+    println(input)
+
+    input = split(input, '\n')
+    if isempty(input[end])
+        pop!(input)
     end
 
     return input
